sch_htb: Use helpers to read stats in ->dump_stats().
The read of packets/bytes statistics in htb_dump_class_stats() appears
not to be synchronized. htb_dump_class_stats() does not acquire locks
but I'm not sure if the other `bstats' that are read can be modified or
are stable while this callback in invoked.

Add a helper to read the two members while synchronizing against
seqcount_t.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>