net: dev: Make rps_lock() disable interrupts.
Interrupts disabling and in the RPS case locking input_pkt_queue case is split
into local_irq_disable() and optional spin_lock().
This breaks on PREEMPT_RT because the spinlock_t typed lock can not be
acquired with disabled interrupts.
The sections in which the lock is acquired is usually short in a sense that it
is not causing long und unbounded latiencies. One exception is the
skb_flow_limit() invocation which may invoke a BPF program.

By moving local_irq_disable()+spin_lock() into rps_lock(), we can keep
interrupts disabled on !RT kernels and enabled on RT kernels. Without
RPS, the needed synchronisation happens as part of local_bh_disable() on
the local CPU. Since interrupts remain enabled, enqueue_to_backlog()
needs to disable interrupts for ____napi_schedule().

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>