mm/vmalloc: Use raw_cpu_ptr() for vmap_block_queue access.
The per-CPU resource vmap_block_queue is accessed via the
get_cpu_var(). That macro disables preemption and then loads the pointer
from the current CPU.
This doesn't work on PREEMPT_RT because a spinlock_t is later accessed
within the preempt-disable section.

There is no need to disable preemption while accessing the per-CPU
struct vmap_block_queue because the list is protected with a spinlock_t.
The per-CPU struct is also accessed cross-CPU in
purge_fragmented_blocks().
It is possible that by using raw_cpu_ptr() the code migrates to another
CPU and uses struct from another CPU. This is fine because the list
locked and the locked section is very short.

Use raw_cpu_ptr() to access vmap_block_queue.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Link: https://lore.kernel.org/r/YnKx3duAB53P7ojN@linutronix.de