printk: Don't invoke the BLK-console with disabled interrupts.
On PREEMPT_RT the BKL-console is always invoked from the thread, never
from the caller. Disabling interrupts before invoking the console is
broken because the BLK-console is using sleeping locks which must not be
acquired with disabled interrupts.

Replace the disabling of interrupts with a local_lock_t inside the
structure which is used to defer parallel printks on the same CPU.

With this change the BKL-console provides output again without rising to
deadlock. Without atomic consoles, any BUG()/ panic() remains silent.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>