printk: Add kthread for all legacy consoles
The write callback of legacy consoles make use of spinlocks. This is
not permitted with PREEMPT_RT in many contexts.

Create a new kthread to handling printing of all the legacy consoles
(and nbcon consoles if boot consoles are registered).

Since the consoles are printing in a task context, it is no longer
appropriate to support the legacy handover mechanism.

These changes exist only for CONFIG_PREEMPT_RT.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>