printk: nbcon: Add context to console_is_usable()
The nbcon consoles have 2 callbacks to be used for different
contexts. In order to determine if an nbcon console is usable,
console_is_usable() needs to know if it is a context that
requires atomic support or is sufficient with threading
support.

Add an extra parameter @use_atomic to denote this.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>