printk/ringbuffer: Extend the sequence number properly on 32bit.
The sequence number on 32bit platform is extendend to a full 64bit
variable from a 32bit value and a 64bit previous record.
The formala as-is is okay but it relies on the right part to be negative
which is not the case if the result is a unsigned variable.

Cast the result to a signed variable.

Reported-by: Francesco Dolcini <francesco@dolcini.it>
Reported-by: kernel test robot <oliver.sang@intel.com>
Closes: https://lore.kernel.org/oe-lkp/202311171611.78d41dbe-oliver.sang@intel.com
Reported-by: kernel test robot <oliver.sang@intel.com>
Closes: https://lore.kernel.org/oe-lkp/202311161555.3ee16fc9-oliver.sang@intel.com
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>